Use copyState rather than loadSnapshot twice.
Calling loadSnapshot twice prevents the "easy" and "optimal"
implementations when the object is already trivially representable as
JSON, since it results in two state object references to the same
object, and thus we end up with updates being duplicated in the
predicted state.

By using copyState instead, the optimal implementation works as
expected.